Interfaces for ships and events
We'll use domain events to decouple the domain with the implementation. Ships will fire off an attack event when they are performing an attack against a specific enemy. This can then get handled at some other level. First pass, the ship is handling 'taking a hit', although this too may move to a 'damage engine' or something along those lines.