Handle shield and structure hits separately
The 'attack' pipeline will be orchestrated by some sort of attack handler. This handler will interact with the ship to apply shield and structure hits to it. First the ship will run a shield hit to modify the attack, then the remaining attack 'damage' will be applied to a structure hit. This gives ships the ability to 'absorb' damage with a shield-like ability even if they don't have normal shields, such as damper field.