Implemented a weapon class
The weapon will handle the attack event and use it to interact with the ship shield and structure hit actions. This gives the weapon the ability to apply special abilities when attacking the shield and/or structure. (We may want to consider breaking this up into smaller functions that the child classes can override.